+++ tests/testthat/test-mahalanobis_print.R
@@ -33,8 +33,12 @@ md_fit1
 md_predictors
 
 print(md_rerun, first = 7)
+print(md_rerun, first = NULL)
 print(md_fit, first = 3, sort = FALSE)
 print(md_fit, first = 3, sort = TRUE, decreasing = FALSE)
 print(md_fit1, first = 3, sort = TRUE, decreasing = FALSE)
 print(md_predictors, first = 10)
+
+print(md_fit1)
+unclass(md_fit1)