+++ NEWS.md
@@ -1,4 +1,4 @@
-# semfindr 0.1.4.15
+# semfindr 0.1.4.23
 
 - Fixed the pkgdown site. (0.1.4.1)
 - Check whether a model has at least one equality constraint
@@ -27,6 +27,27 @@
 - Added an article for multiple-group models. (0.1.4.15)
 - Restructured articles and references in the `pgkdown`
   websites. (0.1.4.15)
+- Changed the default number of cases to printed to 10
+  for the print methods of `est_change`, `influence_stat`,
+  `md_semfindr`, and `fit_measures_change` objects. (0.1.4.16)
+- In `print.lavaan_rerun()`, keep the format of lavaan
+  warnings. (0.1.4.17)
+- In `print.est_change()`, cases are still sorted by
+  gCD even if it is NA for some cases. (0.1.4.18)
+- In `print.md_semfindr()`, cases are still sorted by
+  even if MD is NA for some cases. (0.1.4.18)
+- Revised `print.est_change()` to make its interface
+  similar to that of `print.fit_measures_change()`. (0.1.4.19)
+- Revised plot functions of `est_change()` and friends. They
+  now work properly for the output of `influence_stat()`.
+  (0.1.4.20)
+- Remark that `fit_measures_change_approx()` only supports
+  selected fit measures. (0.1.4.21)
+- Add `cutoff_change` and `largest_change` to
+  `est_change_gcd_plot()`. (0.1.4.22)
+- Updated a few tests on parameters which are fixed in
+  the model but their standardized versions are free.
+  (0.1.4.23)
 
 # semfindr 0.1.4